chore: Disable Windows build and releases
As Windows app requires code signing with expensive certificates,
we will only build and release official linux and macos versions
of mini video me.

The windows app will still be maintained but the user will have to
build it manually by downloading the source code and running the
app with Node.js.